1st setup for MacOS app bundle

+++ openhantek/src/docks/VoltageDock.cpp
@@ -76,7 +76,7 @@ VoltageDock::VoltageDock(DsoSettingsScope *scope, const Dso::ControlSpecificatio
         if (channel < spec->channels) {
             dockLayout->addWidget( b.attnCheckBox, row, 1 ) ;
             dockLayout->addWidget( b.miscComboBox, row++, 2 ) ;
-            if ( scope->voltage[channel].couplingOrMathIndex < couplingStrings.size() )
+            if ( (int)scope->voltage[channel].couplingOrMathIndex < couplingStrings.size() )
                 setCoupling(channel, scope->voltage[channel].couplingOrMathIndex);
         } else {
             dockLayout->addWidget( b.miscComboBox, row++, 1, 1, 2 );
